Move clear_memory_resource_cache to use site
This fixture is only used in one place.
